Moved validation to own file
